refactored dockr lib
